+++ lua53/src/main/java/party/iroiro/luajava/lua53/Lua53Natives.java
@@ -219,7 +219,6 @@ public class Lua53Natives extends LuaNative {
      *
      * <p>
      * The value of <code>op</code> must be one of the following constants:
-     *
      * </p>
      *
      * <ul>
@@ -355,7 +354,6 @@ public class Lua53Natives extends LuaNative {
      *
      * <p>
      * The value of <code>op</code> must be one of the following constants:
-     *
      * </p>
      *
      * <ul>
@@ -525,7 +523,6 @@ public class Lua53Natives extends LuaNative {
      * <p>
      * This function performs several tasks,
      * according to the value of the parameter <code>what</code>:
-     *
      * </p>
      *
      * <ul>
@@ -790,7 +787,7 @@ public class Lua53Natives extends LuaNative {
      * @param i i
      * @return see description
      */
-    protected native int lua_geti(long ptr, int index, int i); /*
+    protected native int lua_geti(long ptr, int index, long i); /*
         lua_State * L = (lua_State *) ptr;
 
         jint returnValueReceiver = (jint) lua_geti((lua_State *) L, (int) index, (lua_Integer) i);
@@ -824,7 +821,7 @@ public class Lua53Natives extends LuaNative {
      * @param index the stack position of the element
      * @param i i
      */
-    protected native void luaJ_geti(long ptr, int index, int i); /*
+    protected native void luaJ_geti(long ptr, int index, long i); /*
         lua_State * L = (lua_State *) ptr;
 
         lua_geti((lua_State *) L, (int) index, (lua_Integer) i);
@@ -1534,7 +1531,7 @@ public class Lua53Natives extends LuaNative {
      * @param size size
      * @return see description
      */
-    protected native long lua_newuserdata(long ptr, int size); /*
+    protected native long lua_newuserdata(long ptr, long size); /*
         lua_State * L = (lua_State *) ptr;
 
         jlong returnValueReceiver = (jlong) lua_newuserdata((lua_State *) L, (size_t) size);
@@ -1563,7 +1560,6 @@ public class Lua53Natives extends LuaNative {
      *
      * <p>
      * A typical traversal looks like this:
-     *
      * </p>
      *
      * <pre>
@@ -1657,7 +1653,6 @@ public class Lua53Natives extends LuaNative {
      * <p>
      * The <a href="https://www.lua.org/manual/5.3/manual.html#lua_pcall"><code>lua_pcall</code></a> function returns one of the following constants
      * (defined in <code>lua.h</code>):
-     *
      * </p>
      *
      * <ul>
@@ -1758,7 +1753,6 @@ public class Lua53Natives extends LuaNative {
      * <p>
      * The <a href="https://www.lua.org/manual/5.3/manual.html#lua_pcall"><code>lua_pcall</code></a> function returns one of the following constants
      * (defined in <code>lua.h</code>):
-     *
      * </p>
      *
      * <ul>
@@ -1895,9 +1889,9 @@ public class Lua53Natives extends LuaNative {
      * </p>
      *
      * @param ptr the <code>lua_State*</code> pointer
-     * @param n the number of elements
+     * @param n the number / the number of elements
      */
-    protected native void lua_pushinteger(long ptr, int n); /*
+    protected native void lua_pushinteger(long ptr, long n); /*
         lua_State * L = (lua_State *) ptr;
 
         lua_pushinteger((lua_State *) L, (lua_Integer) n);
@@ -1979,7 +1973,7 @@ public class Lua53Natives extends LuaNative {
      * </p>
      *
      * @param ptr the <code>lua_State*</code> pointer
-     * @param n the number of elements
+     * @param n the number / the number of elements
      */
     protected native void lua_pushnumber(long ptr, double n); /*
         lua_State * L = (lua_State *) ptr;
@@ -2227,10 +2221,10 @@ public class Lua53Natives extends LuaNative {
      *
      * @param ptr the <code>lua_State*</code> pointer
      * @param index the stack position of the element
-     * @param n the number of elements
+     * @param n the number / the number of elements
      * @return see description
      */
-    protected native int lua_rawgeti(long ptr, int index, int n); /*
+    protected native int lua_rawgeti(long ptr, int index, long n); /*
         lua_State * L = (lua_State *) ptr;
 
         jint returnValueReceiver = (jint) lua_rawgeti((lua_State *) L, (int) index, (lua_Integer) n);
@@ -2262,7 +2256,7 @@ public class Lua53Natives extends LuaNative {
      *
      * @param ptr the <code>lua_State*</code> pointer
      * @param index the stack position of the element
-     * @param n the number of elements
+     * @param n the number / the number of elements
      */
     protected native void luaJ_rawgeti(long ptr, int index, int n); /*
         lua_State * L = (lua_State *) ptr;
@@ -2332,10 +2326,10 @@ public class Lua53Natives extends LuaNative {
      * @param index the stack position of the element
      * @return see description
      */
-    protected native int lua_rawlen(long ptr, int index); /*
+    protected native long lua_rawlen(long ptr, int index); /*
         lua_State * L = (lua_State *) ptr;
 
-        jint returnValueReceiver = (jint) lua_rawlen((lua_State *) L, (int) index);
+        jlong returnValueReceiver = (jlong) lua_rawlen((lua_State *) L, (int) index);
         return returnValueReceiver;
     */
 
@@ -2675,9 +2669,9 @@ public class Lua53Natives extends LuaNative {
      *
      * @param ptr the <code>lua_State*</code> pointer
      * @param index the stack position of the element
-     * @param n the number of elements
+     * @param n the number / the number of elements
      */
-    protected native void lua_seti(long ptr, int index, int n); /*
+    protected native void lua_seti(long ptr, int index, long n); /*
         lua_State * L = (lua_State *) ptr;
 
         lua_seti((lua_State *) L, (int) index, (lua_Integer) n);
@@ -2894,10 +2888,10 @@ public class Lua53Natives extends LuaNative {
      * @param s the string
      * @return see description
      */
-    protected native int lua_stringtonumber(long ptr, String s); /*
+    protected native long lua_stringtonumber(long ptr, String s); /*
         lua_State * L = (lua_State *) ptr;
 
-        jint returnValueReceiver = (jint) lua_stringtonumber((lua_State *) L, (const char *) s);
+        jlong returnValueReceiver = (jlong) lua_stringtonumber((lua_State *) L, (const char *) s);
         return returnValueReceiver;
     */
 
@@ -2955,10 +2949,10 @@ public class Lua53Natives extends LuaNative {
      * @param index the stack position of the element
      * @return see description
      */
-    protected native int lua_tointeger(long ptr, int index); /*
+    protected native long lua_tointeger(long ptr, int index); /*
         lua_State * L = (lua_State *) ptr;
 
-        jint returnValueReceiver = (jint) lua_tointeger((lua_State *) L, (int) index);
+        jlong returnValueReceiver = (jlong) lua_tointeger((lua_State *) L, (int) index);
         return returnValueReceiver;
     */
 
@@ -2993,10 +2987,10 @@ public class Lua53Natives extends LuaNative {
      * @param isnum pointer to a boolean to be assigned
      * @return see description
      */
-    protected native int lua_tointegerx(long ptr, int index, long isnum); /*
+    protected native long lua_tointegerx(long ptr, int index, long isnum); /*
         lua_State * L = (lua_State *) ptr;
 
-        jint returnValueReceiver = (jint) lua_tointegerx((lua_State *) L, (int) index, (int *) isnum);
+        jlong returnValueReceiver = (jlong) lua_tointegerx((lua_State *) L, (int) index, (int *) isnum);
         return returnValueReceiver;
     */
 
@@ -3638,7 +3632,6 @@ public class Lua53Natives extends LuaNative {
      * <p>
      * Loads and runs the given string.
      * It is defined as the following macro:
-     *
      * </p>
      *
      * <pre>
@@ -3902,10 +3895,10 @@ public class Lua53Natives extends LuaNative {
      * @param index the stack position of the element
      * @return see description
      */
-    protected native int luaL_len(long ptr, int index); /*
+    protected native long luaL_len(long ptr, int index); /*
         lua_State * L = (lua_State *) ptr;
 
-        jint returnValueReceiver = (jint) luaL_len((lua_State *) L, (int) index);
+        jlong returnValueReceiver = (jlong) luaL_len((lua_State *) L, (int) index);
         return returnValueReceiver;
     */
 
@@ -4334,7 +4327,6 @@ public class Lua53Natives extends LuaNative {
      * Pushes onto the stack a string identifying the current position
      * of the control at level <code>lvl</code> in the call stack.
      * Typically this string has the following format:
-     *
      * </p>
      *
      * <pre>
