+++ lua54/src/main/java/party/iroiro/luajava/lua54/Lua54Natives.java
@@ -227,7 +227,6 @@ public class Lua54Natives extends LuaNative {
      *
      * <p>
      * The value of <code>op</code> must be one of the following constants:
-     *
      * </p>
      *
      * <ul>
@@ -448,7 +447,6 @@ public class Lua54Natives extends LuaNative {
      *
      * <p>
      * The value of <code>op</code> must be one of the following constants:
-     *
      * </p>
      *
      * <ul>
@@ -789,7 +787,7 @@ public class Lua54Natives extends LuaNative {
      * @param i i
      * @return see description
      */
-    protected native int lua_geti(long ptr, int index, int i); /*
+    protected native int lua_geti(long ptr, int index, long i); /*
         lua_State * L = (lua_State *) ptr;
 
         jint returnValueReceiver = (jint) lua_geti((lua_State *) L, (int) index, (lua_Integer) i);
@@ -823,7 +821,7 @@ public class Lua54Natives extends LuaNative {
      * @param index the stack position of the element
      * @param i i
      */
-    protected native void luaJ_geti(long ptr, int index, int i); /*
+    protected native void luaJ_geti(long ptr, int index, long i); /*
         lua_State * L = (lua_State *) ptr;
 
         lua_geti((lua_State *) L, (int) index, (lua_Integer) i);
@@ -1545,7 +1543,7 @@ public class Lua54Natives extends LuaNative {
      * @param nuvalue number of associated Lua values (user values)
      * @return see description
      */
-    protected native long lua_newuserdatauv(long ptr, int size, int nuvalue); /*
+    protected native long lua_newuserdatauv(long ptr, long size, int nuvalue); /*
         lua_State * L = (lua_State *) ptr;
 
         jlong returnValueReceiver = (jlong) lua_newuserdatauv((lua_State *) L, (size_t) size, (int) nuvalue);
@@ -1574,7 +1572,6 @@ public class Lua54Natives extends LuaNative {
      *
      * <p>
      * A typical table traversal looks like this:
-     *
      * </p>
      *
      * <pre>
@@ -1843,9 +1840,9 @@ public class Lua54Natives extends LuaNative {
      * </p>
      *
      * @param ptr the <code>lua_State*</code> pointer
-     * @param n the number of elements
+     * @param n the number / the number of elements
      */
-    protected native void lua_pushinteger(long ptr, int n); /*
+    protected native void lua_pushinteger(long ptr, long n); /*
         lua_State * L = (lua_State *) ptr;
 
         lua_pushinteger((lua_State *) L, (lua_Integer) n);
@@ -1927,7 +1924,7 @@ public class Lua54Natives extends LuaNative {
      * </p>
      *
      * @param ptr the <code>lua_State*</code> pointer
-     * @param n the number of elements
+     * @param n the number / the number of elements
      */
     protected native void lua_pushnumber(long ptr, double n); /*
         lua_State * L = (lua_State *) ptr;
@@ -2177,10 +2174,10 @@ public class Lua54Natives extends LuaNative {
      *
      * @param ptr the <code>lua_State*</code> pointer
      * @param index the stack position of the element
-     * @param n the number of elements
+     * @param n the number / the number of elements
      * @return see description
      */
-    protected native int lua_rawgeti(long ptr, int index, int n); /*
+    protected native int lua_rawgeti(long ptr, int index, long n); /*
         lua_State * L = (lua_State *) ptr;
 
         jint returnValueReceiver = (jint) lua_rawgeti((lua_State *) L, (int) index, (lua_Integer) n);
@@ -2212,7 +2209,7 @@ public class Lua54Natives extends LuaNative {
      *
      * @param ptr the <code>lua_State*</code> pointer
      * @param index the stack position of the element
-     * @param n the number of elements
+     * @param n the number / the number of elements
      */
     protected native void luaJ_rawgeti(long ptr, int index, int n); /*
         lua_State * L = (lua_State *) ptr;
@@ -2654,9 +2651,9 @@ public class Lua54Natives extends LuaNative {
      *
      * @param ptr the <code>lua_State*</code> pointer
      * @param index the stack position of the element
-     * @param n the number of elements
+     * @param n the number / the number of elements
      */
-    protected native void lua_seti(long ptr, int index, int n); /*
+    protected native void lua_seti(long ptr, int index, long n); /*
         lua_State * L = (lua_State *) ptr;
 
         lua_seti((lua_State *) L, (int) index, (lua_Integer) n);
@@ -2897,10 +2894,10 @@ public class Lua54Natives extends LuaNative {
      * @param s the string
      * @return see description
      */
-    protected native int lua_stringtonumber(long ptr, String s); /*
+    protected native long lua_stringtonumber(long ptr, String s); /*
         lua_State * L = (lua_State *) ptr;
 
-        jint returnValueReceiver = (jint) lua_stringtonumber((lua_State *) L, (const char *) s);
+        jlong returnValueReceiver = (jlong) lua_stringtonumber((lua_State *) L, (const char *) s);
         return returnValueReceiver;
     */
 
@@ -3009,10 +3006,10 @@ public class Lua54Natives extends LuaNative {
      * @param index the stack position of the element
      * @return see description
      */
-    protected native int lua_tointeger(long ptr, int index); /*
+    protected native long lua_tointeger(long ptr, int index); /*
         lua_State * L = (lua_State *) ptr;
 
-        jint returnValueReceiver = (jint) lua_tointeger((lua_State *) L, (int) index);
+        jlong returnValueReceiver = (jlong) lua_tointeger((lua_State *) L, (int) index);
         return returnValueReceiver;
     */
 
@@ -3047,10 +3044,10 @@ public class Lua54Natives extends LuaNative {
      * @param isnum pointer to a boolean to be assigned
      * @return see description
      */
-    protected native int lua_tointegerx(long ptr, int index, long isnum); /*
+    protected native long lua_tointegerx(long ptr, int index, long isnum); /*
         lua_State * L = (lua_State *) ptr;
 
-        jint returnValueReceiver = (jint) lua_tointegerx((lua_State *) L, (int) index, (int *) isnum);
+        jlong returnValueReceiver = (jlong) lua_tointegerx((lua_State *) L, (int) index, (int *) isnum);
         return returnValueReceiver;
     */
 
@@ -3713,7 +3710,6 @@ public class Lua54Natives extends LuaNative {
      * <p>
      * Loads and runs the given string.
      * It is defined as the following macro:
-     *
      * </p>
      *
      * <pre>
@@ -3977,10 +3973,10 @@ public class Lua54Natives extends LuaNative {
      * @param index the stack position of the element
      * @return see description
      */
-    protected native int luaL_len(long ptr, int index); /*
+    protected native long luaL_len(long ptr, int index); /*
         lua_State * L = (lua_State *) ptr;
 
-        jint returnValueReceiver = (jint) luaL_len((lua_State *) L, (int) index);
+        jlong returnValueReceiver = (jlong) luaL_len((lua_State *) L, (int) index);
         return returnValueReceiver;
     */
 
@@ -4432,7 +4428,6 @@ public class Lua54Natives extends LuaNative {
      * Pushes onto the stack a string identifying the current position
      * of the control at level <code>lvl</code> in the call stack.
      * Typically this string has the following format:
-     *
      * </p>
      *
      * <pre>
